requirements.txt
# Core dependencies
python-dotenv
langchain
streamlit
pandas
numpy
faiss-cpu
groq
sentence-transformers

# Data processing
openpyxl
python-pptx
pypdf2

# Utilities
tqdm
plotly
selenium
webdriver_manager

matplotlib
scikit-learn